feat: update duration type to Double and enhance song detail loading logic

+++ app/src/main/java/com/jgm90/cloudmusic/core/innertube/YouTubeRepository.kt
@@ -344,7 +344,7 @@ class YouTubeRepository @Inject constructor(
         val trackName: String?,
         val artistName: String?,
         val albumName: String?,
-        val duration: Int?,
+        val duration: Double?,
         val instrumental: Boolean?,
         val plainLyrics: String?,
         val syncedLyrics: String?
